requirements.txt
streamlit
torch
torchvision
transformers
faiss-cpu==1.10.0
pillow
numpy
pandas
sentence-transformers
datasets
python-dotenv
google-generativeai
fastapi
uvicorn
sqlalchemy
pyjwt
passlib
bcrypt
python-multipart
